Pass the annotations set on a queryset to the specific() queryset
In case you want to execute a query like:

  root_page.get_children().annotate(something=Value('something')).specific()

The annotation get's lost in the specific() call. By reading all the
annotations set on the queryset and setting them on the final queryset
we can keep this intact.

Note that this commit is more to demonstrate the approach. Unittests etc
will be added when this turns out to be a valid approach